backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4
requests==2.31.0
resend==0.6.0
web3==6.11.3
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
